requirements.txt
Django>=4.0,<5.0
google-generativeai
python-dotenv
django-environ
